add comment post and delete comment successfully

+++ controllers/api/v1/post_api.js
@@ -26,8 +26,8 @@ module.exports.destroy= async function(req,res){
         let post=await Post.findById(req.params.id);
 
 
-        // if(post.user==req.user.id){
-            post.remove();
+         if(post.user==req.user.id){
+             post.remove();//are you there
             
             await Comment.deleteMany({post:req.params.id});
 
@@ -38,11 +38,12 @@ module.exports.destroy= async function(req,res){
     
     
     
-        //    }  else{
-        //     req.flash('error','You cannot delete this post!');
-
-        //        return res.redirect('back');
-        //    }
+           }  else{
+   
+               return res.json(401,{
+                   message:"You cannot delete this message"
+               });
+           }
         
     
     }catch(err){
